adds search bar component
